requirements.txt
pip
pillow
numpy
spidev
gpiozero
dotenv